+++ public/js/obra.js
@@ -41,21 +41,47 @@ $(document).ready(function(){
 
     //delete obra and remove it from list
     $('#obra-list').on('click', '.delete-obra',function(){
+        $.ajaxSetup({
+            headers: {
+                'X-CSRF-TOKEN': $('meta[name="_token"]').attr('content')
+            }
+        })
         var obra_id = $(this).val();
 
-        $.ajax({
-
-            type: "DELETE",
-            url: url + '/' + obra_id,
-            success: function (data) {
-                console.log(data);
-
-                $("#obra" + obra_id).remove();
+        event.preventDefault();
+        swal({
+                title: "Estas seguro?",
+                text: "Se eliminara permanentemente de la base de datos!",
+                type: "warning",
+                showCancelButton: true,
+                confirmButtonColor: "#DD6B55",
+                confirmButtonText: "SI, borrar!",
+                cancelButtonText: "NO, cancelar!",
+                closeOnConfirm: true,
+                closeOnCancel: false
             },
-            error: function (data) {
-                console.log('Error:', data);
-            }
-        });
+            function(isConfirm){
+                if (isConfirm) {
+                    swal("Borrado!", "Fue borrado exitosamente.", "success");
+                    $.ajax({
+
+                        type: "DELETE",
+                        url: url + '/' + obra_id,
+                        success: function (data) {
+
+                            console.log(data);
+
+                            $("#obra" + obra_id).remove();
+                            window.location.reload();
+                        },
+                        error: function (data) {
+                            console.log('Error:', data);
+                        }
+                    });
+                } else {
+                    swal("Cancelado", "El texto no fue borrado :)", "error");
+                }
+            });
     });
 
     //create new obra / update existing obra
@@ -114,6 +140,7 @@ $(document).ready(function(){
                 $('#frmobra').trigger("reset");
 
                 $('#myModal').modal('hide')
+                window.location.reload();
             },
             error: function (data) {
                 console.log('Error:', data);
